Apply suggested corrections
Revert darwin module changes
Revert change to ffi binding in fc::*
Revert changes to std import nesting
Revert addition of self prefix within fc::pattern
Remove superfluous brackets in fc
