+++ font/src/ft/fc/font_set.rs
@@ -11,9 +11,10 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-use std::{ops::Deref, ptr::NonNull};
+use std::ops::Deref;
+use std::ptr::NonNull;
 
-use fontconfig::fontconfig::{FcFontSet, FcFontSetDestroy, FcFontSetList};
+use super::ffi::{FcFontSet, FcFontSetDestroy, FcFontSetList};
 use foreign_types::{foreign_type, ForeignType, ForeignTypeRef};
 use log::trace;
